Crate spin_compose

Enums

ComposeError

Represents an error that can occur when composing dependencies.

Traits

ComponentSourceLoader

This trait is used to load component source code from a locked component source across various embdeddings.

Functions

compose

Composes a Spin AppComponent using the dependencies specified in the component’s dependencies section.