spin_compose

Trait ComponentSourceLoader

pub trait ComponentSourceLoader {
    type Component: ComponentLike<Dependency = Self::Dependency>;
    type Dependency: DependencyLike;

    // Required methods
    fn load_component_source<'life0, 'life1, 'async_trait>(
        &'life0 self,
        source: &'life1 Self::Component,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn load_dependency_source<'life0, 'life1, 'async_trait>(
        &'life0 self,
        source: &'life1 Self::Dependency,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

This trait is used to load component source code from a locked component source across various embdeddings.

Required Associated Types

type Component: ComponentLike<Dependency = Self::Dependency>

type Dependency: DependencyLike

Required Methods

fn load_component_source<'life0, 'life1, 'async_trait>( &'life0 self, source: &'life1 Self::Component, ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn load_dependency_source<'life0, 'life1, 'async_trait>( &'life0 self, source: &'life1 Self::Dependency, ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,