Crate spin_compose

Structs

ComponentSourceLoaderFs

A ComponentSourceLoader that loads component sources from the filesystem.

Enums

ComposeError

Represents an error that can occur when composing dependencies.

InheritConfiguration

Traits

ComponentLike

A Spin component. This abstracts over the list of dependencies for the component. The abstraction allows both manifest and lockfile types to participate in composition.

ComponentSourceLoader

This trait is used to load component source code from a locked component source across various embdeddings.

DependencyLike

A Spin component dependency. This abstracts over the metadata associated with the dependency. The abstraction allows both manifest and lockfile types to participate in composition.

Functions

compose

Composes a Spin AppComponent using the dependencies specified in the component’s dependencies section.